% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqmClip.R
\name{aqmClip}
\alias{aqmClip}
\title{Title}
\usage{
aqmClip(
  mpaShp,
  buff,
  siteName,
  nowFut = c("now", "fut"),
  corr = c(NA, "lonFx"),
  datFolder
)
}
\arguments{
\item{mpaShp}{shp file of the existing or proposed MPA}

\item{buff}{distance around mpa shape file for which analysis is conducted (to calculate extent box, "ploArea"). In decimal degrees (numeric)}

\item{siteName}{character string of site name (no spaces) for file name creation}

\item{nowFut}{one of "now" or "fut" to indicate which dataset to pull from: now for current ranges, fut for 2050 projections}

\item{corr}{Indicates whether longitude "lonFx" needs to be corrected for crossing the 180 meridian, if no use NA}

\item{datFolder}{file path to where aqm data are stored}
}
\value{
writes a spatial features dataframe with aquamaps data from the indicated dataset (now/fut) for the defined plot area
}
\description{
Title
}
\examples{

aqmClip(mpaShp=revMPA,buff=10,siteName="rev",datFolder="../../../ccAssmts/data")
}
