% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqmTurnDiss.R
\name{aqmTurnDiss}
\alias{aqmTurnDiss}
\title{Calculate turnover, nestedness or total dissimilarity}
\usage{
aqmTurnDiss(
  mpaShp,
  siteName,
  buff,
  thresh = 0.4,
  dissCalc = c("turnover", "nestedness", "dissimilarity"),
  dissIndx = c("sorensen", "jaccard"),
  corr = c(NA, "lonFx")
)
}
\arguments{
\item{mpaShp}{shp file of the existing or proposed MPA}

\item{siteName}{character string of site name (no spaces) for file name creation (should match aqmClip value)}

\item{buff}{distance around mpa shape file for which analysis is conducted (to calculate extent box, "ploArea"). In decimal degrees (numeric)}

\item{thresh}{value from 0-1 to indicate presence threshold to filter data (usually one of 0.4, 0.5, 0.6, 0.7,default is 0.4)}

\item{dissCalc}{one of "turnover","nestedness","dissimilarity" indicating which metric of dissimilarity to plot. All are calculated based on the dissIndex each run}

\item{dissIndx}{one of "sorensen" or "jaccard" to indicate which metric of turnover and dissimilarity to use}

\item{corr}{indicates whether longitude "lonFx" needs to be corrected for crossing the 180 meridian, if no use NA}
}
\value{
ggplot map object with 0.5 degree raster of dissimilarity, with MPA shapefile, plot area, eez, and surrounding countries plotted
}
\description{
Calculate turnover, nestedness or total dissimilarity
}
\examples{
aqmTurnDiss(mpaShp=revMPA,siteName="rev",buff=10,thresh=0.6,dissCalc="turnover",dissIndex="jaccard",coor=NA)
}
