% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSppRich.R
\name{dSppRich}
\alias{dSppRich}
\title{Changes to species richness}
\usage{
dSppRich(mpaShp, siteName, buff, thresh = 0.4, corr = c(NA, "lonFx"))
}
\arguments{
\item{mpaShp}{shp file of the existing or proposed MPA, creates different layers using mpaShpBbBase function}

\item{siteName}{character string of site name (no spaces) for file name creation (should match aqmClip value)}

\item{buff}{distance around mpa shape file for which analysis is conducted (to calculate extent box, "ploArea"). In decimal degrees (numeric)}

\item{thresh}{value from 0-1 to indicate presence threshold to filter data (usually one of 0.4, 0.5, 0.6, 0.7,default is 0.4)}

\item{corr}{indicates whether longitude "lonFx" needs to be corrected for crossing the 180 meridian, if no use NA}
}
\value{
map of 0.5 degree raster of changes +/- in species richness
}
\description{
Takes the output data from aqmClip and calculates the absolute change in species richness for the indicates region
}
\examples{
dSppRich(revMPA,siteName="rev",buff=10,thresh=0.6,corr=NA)
}
