% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extMap.R
\name{extMap}
\alias{extMap}
\title{Extent Map}
\usage{
extMap(mpaShp, buff, zoomLat, zoomLon, corr = c(NA, "lonFx"))
}
\arguments{
\item{mpaShp}{shp file of the existing or proposed MPA, creates different layers using mpaShpBbBase function}

\item{buff}{distance around mpa shape file for which analysis is conducted (to calculate extent box, "ploArea"). In decimal degrees (numeric)}

\item{zoomLat}{latitudinal buffer for the entire extent map in decimal degrees (size depends on context extent map scale)}

\item{zoomLon}{longitudinal buffer for the entire extent map in decimal degrees (size depends on context extent map scale)}

\item{corr}{indicates whether longitude "lonFx" needs to be corrected for crossing the 180 meridian, if no use NA}
}
\value{
extent map of the plot area around the mpaShp based on buff
}
\description{
takes a MPA shapefile and creates an extent map showing the area assessed based on the established plot area
}
\examples{
extMap(mpaShp=revMPA, buff=7,zoomLat=15,zoomLon=10,corr=NA)
}
