% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extMap.R
\name{extMapFunc}
\alias{extMapFunc}
\title{Title}
\usage{
extMapFunc(mpaShp, buff, zoomLat, zoomLon, corr = c(NA, "lonFx"))
}
\arguments{
\item{mpaShp}{shp file of the existing or proposed MPA}

\item{buff}{distance around mpa shape file for which analysis is conducted (to calculate extent box, "ploArea"). In decimal degrees (numeric)}

\item{zoomLat}{latitudinal buffer for the entire extent map in decimal degrees (size depends on context extent map scale)}

\item{zoomLon}{longitudinal buffer for the entire extent map in decimal degrees (size depends on context extent map scale)}

\item{corr}{indicates whether longitude "lonFx" needs to be corrected for crossing the 180 meridian, if no use NA}
}
\value{
extent map of the plot area around the mpaShp based on buff
}
\description{
Title
}
