% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhwMapYrs.R
\name{mhwMapYrs}
\alias{mhwMapYrs}
\title{Plot marine heatwave data spatially and temporally}
\usage{
mhwMapYrs(
  mpaShp,
  mhwMetric,
  ssps,
  buff,
  YEAR = 2050,
  corr = c(NA, "lonFx"),
  datFolder
)
}
\arguments{
\item{mpaShp}{shp file of the existing or proposed MPA}

\item{mhwMetric}{one of "dur" (mean duration), "intns" (maximal intensity), "sum" (number of MHW days per year)}

\item{ssps}{one of 126 or 585 indicating the projection scenario to look at}

\item{buff}{distance around mpa shape file for which analysis is conducted (to calculate extent box, "ploArea"). In decimal degrees (numeric)}

\item{YEAR}{year in the future for analysis from 2020 - 2100, default is 2050}

\item{corr}{indicates whether longitude (lonFx) needs to be corrected for crossing the 180 meridian, if no use NA}

\item{datFolder}{file path to where mhw data are stored}
}
\value{
two plots: 1) raster map of indicated MHW metric within indicated region, 2) plot of MHW metric over time
}
\description{
Plot marine heatwave data spatially and temporally
}
\examples{
mhwMpaYrs(mpaShp=revMPA,mhwMetric="dur",ssps=585,buff=10,year=2075,corr=NA,datFolder="../../../ccAssmts/data")
}
