% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpaShpBbBase.R
\name{mpaShpBbBase}
\alias{mpaShpBbBase}
\title{Title}
\usage{
mpaShpBbBase(mpaShp, buff, corr = c(NA, "lonFx"))
}
\arguments{
\item{mpaShp}{shp file of the existing or proposed MPA}

\item{buff}{distance around mpa shape file for which analysis is conducted (to calculate extent box, "ploArea"). In decimal degrees (numeric)}

\item{corr}{indicates whether longitude "lonFx" needs to be corrected for crossing the 180 meridian, if no use NA}
}
\value{
list of 3 items: mpaShp projected to WGS84 and corrected to 360 if needed: corr="lonFx", bounding box for the plot area, base map of land within the plot area
}
\description{
Title
}
