% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrEEZ}
\alias{mrEEZ}
\title{Marine Regions EEZ Boundaries Data}
\format{
A spatial feature dataset with several identifying columns, and line (MULTISTRING) geometries. Contains 27 variables of 2215 objects.
}
\usage{
mrEEZ
}
\description{
Spatial features (sf) data frame with the Exclusive Economic Zones for all countries around the world for plotting purposes. Data are stored as lines.
}
\keyword{datasets}
