% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revMPA}
\alias{revMPA}
\title{Revillagigedo MPA Spatial data}
\format{
A spatial feature dataset with several identifying columns, and line (MULTISTRING) geometries. Contains 31 variables of 2 objects.
}
\usage{
revMPA
}
\description{
Spatial features (sf) data frame of the Revillagigedo MPA off the Pacific coast of Mexico.
}
\keyword{datasets}
