% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaaCmip6.R
\name{noaaCMIP6}
\alias{noaaCMIP6}
\title{Plot NOAA CMIP6 data}
\usage{
noaaCMIP6(
  mpaShp,
  buff,
  futYr,
  var = c("tos", "sos", "o2"),
  corr = c(NA, "lonFx")
)
}
\arguments{
\item{mpaShp}{shp file of the existing or proposed MPA}

\item{buff}{distance around mpa shape file for which analysis is conducted (to calculate extent box, "ploArea"). In decimal degrees (numeric)}

\item{futYr}{year to capture variable projection. Data range from 2015-2100}

\item{var}{one of "tos" (SST), "sos" (SSSal), "o2s" ([O2] shallow), "o2d" ([O2] deep) variables to plot.SST and SSS data are by definition only calculated at one depth (0m) and oxygen concentration is assessed at shallow and deep}

\item{corr}{indicates whether longitude "lonFx" needs to be corrected for crossing the 180 meridian, if no use NA}
}
\value{
ggplot object of variable projections or, for oxygen, list of 2 ggplot objects of projected oxygen concentations at 2 depths
}
\description{
Pulls NOAA data from author's private drive folder and plots up variable (temperature (tos), salinity (sos), oxygen at 2 depths (o2)) projections for the assessment area indicated
}
\examples{
noaaCMIP6(mpaShp=revMPA,buff=7,var="tos",corr=NA)

}
